frequency: Simplify and fix total/active residency calculations
There are currently two distincts ways of calculating freq residency, whereas we
only need one. The one used for active residency had the advantage of taking
into account the fact that the frequency stays active after the last
cpu_frequency event (by virtue of the fact that it joins the cpu_frequency event
with the cpu_idle event), so use that one.